Create fusing_fp.py
This python scripts can combine the fingerprints for every smiles that targeting the same target (column Target_ID). 
Problem is that in order to combine the fingerprints I had to convert tm.VectorUint to a np.array and in order to run TMAP I have to convert again np.array to tm.VectorUint which, as of today, I don't know how to do.